pyproject.toml
[tool.poetry]
name = "openai-mock-api"
version = "0.1.0"
description = ""
authors = ["LukasGardberg <\u001alukasgardberg2@gmail.com\u001a>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
aioprometheus = "^23.12.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.27.0.post1"
pydantic = "^2.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
